fix: changed all `std::byte`s to `uint8_t`
I probably should have done this from the start but I wanted to do things
the "C++" way. `std::byte` had to be changed because it broke the templated
conversion functions.

+++ src/Localization/LOCFile.cpp
@@ -20,7 +20,7 @@ namespace l4jf::loc {
 
 		// Key Table
 		if(version == 0x2) {
-			useUniqueIds = (reader.ReadByte() > static_cast<std::byte>(0));
+			useUniqueIds = (reader.ReadByte() > 0);
 			uint32_t keyCount =  reader.Read<uint32_t>();
 			
 			keys.resize(keyCount);
